Fix for Soundpool issue in GB
The problem can occur if a sample is started at the same time as
the last AudioTrack callback for a playing sample is called. At
this time, allocateChannel() can be called concurrently with
moveToFront() which can cause an entry in mChannels being used by
moveToFront() to be erased temporarily by allocateChannel().The fix
consists in making sure that the SoundPool mutex is heldwhenever
play(), stop() or done() are called.

CRs fixed: 286433

Change-Id: Ie8f60d8ba5599643b6f48c5dd0cd2cc256e02173

Conflicts:

	media/java/android/media/AudioService.java