Telephony: Send Comprehension Required(CR) flag to Stk app.
CR flag in command details(CD) TLV is always set as true and
sent to Stk app. When CR flag in CD TLV is false in the proactive
command, there is a mismatch of CD TLV data between Stk framework
and Stk app. Terminal Response (TR) is not sent for the proactive
command. Fixed this issue by sending correct CR flag to Stk App.

Change-Id: If7b653cc1e9e1e2d89b16f5903dbb7c7a9f3a605
CRs-Fixed: 292351