frameworks/base: remove app token before finishing the activity
Finishing an activity while its token is active with window
manager, in some cases, causes it to be added to mClosedApps
of window manager. This keeps an active reference to the
surface of the activity. This is meant for animating
activities and not for activities of a crashing process.
Fix is to remove the activity's token before finishing it.

Change-Id: I5a66e34b08ebb55c911811523cba3caa46da09ba
CRs-Fixed: 293155
(cherry picked from commit 0a2853341fedd4bcd596a9f73dfcfcf51b534573)

 On branch 7145