surfaceflinger: Call glFinish in drawWormhole
In full screen videos that are drawn with overlay,
drawWormhole calles glClear for every frame but
there is no eglSwapBuffers or glFinish to flush
the GL command buffers - hence causing a leak in
adreno. Earlier, we worked around this by not calling
drawWormhole when drawing with overlay - however this
caused a side effect with certain 3rd party apps that
use push buffers. Calling a glFinish() after the glClear
in drawWormhole fixes the issue.

Change-Id: I55986b57d87e385f3301929403e6526df63c64fa
CRs-fixed: 293154 280056