Fix build error with AwesomePlayer
It looks like timeUs was renamed to videoTimeUs but the use of the variable
in a LOGV was not updated.  This results in a build error, but only if logging
is enabled.

Someone had fixed this issue previously, but the fix is guarded with an ifdef
for TARGET_OMAP4.  This change removes the ifdef and fixes the issue for all
platforms.