Don't fail check when using YUV420PackedSemiPlanar
OMX_COLOR_FormatYUV420PackedSemiPlanar is used by the OMX codecs on TI81xx, so
we should not fail an assertion here.

This color format was already added to the assertion, but only if TARGET_OMAP4
was defined.  The ifdef was removed rather than addition an additional check
for something like TARGET_TI81XX, since this seems to be a basic sanity-check
assertion, and a failure will still occur later if a platform doesn't support a
specific color format.  Note that OMX_QCOM_COLOR_FormatYVU420SemiPlanar isn't
guarded by an ifdef either for this check, which is clearly specific to a QCOM
device.