Pass OMX buffer header to hardware renderer
pPlatformPrivate is currently used internally by the OMX components (arguably
a spec violation), so we must save and restore this value when calling
the render function, so we can set it to be the OMX buffer header instead.

Passing the OMX buffer header to the hardare renderer is necessary to render
frames properly without a copy.  If the OMX components stop using
pPlatformPrivate at some point, we may be able to find a better place to set
pPlatformPrivate for the renderer.