SharedClient lock failsafe when remote process dies
During Monkey runs, sometimes, a process, holding SharedClient
LOCK, dies while holding that lock. Since SharedClient lock is
of type SHARED, SF and Binder threads cannot acquire it anymore
causing a system hang. Since it seems like posix would not allow
us to unlock a mutex not owned by the caller thread, we are out
of luck and need to ignore that update.
It is assumed to be safe since because if the client dies (process
dies which was the client to SF), so it is safe to ignore all related
pending updates and/or transactions.