frameworks/base: Fix to avoid IndexOutOfBoundsException in am
am throws a IndexOutOfBoundsException in a scenario where a
process, providing content to another process, crashes causing
the client process to be killed. This results in a
IndexOutOfBoundsException.

 In addition, invoking finishActivityLocked for the activities
hosted in a dead process causes duplicate entries in
mCancelledThumbnails, if clients of am are waiting for
thumbnails.

 Fix is not to invoked finishActivityLocked, instead update
am's data structure and send results.

Change-Id: I9a1040b7a021f27c7bc47d1573b3429c7ba28a38