Support for AT +CMAR=<pin_lock_code> & +CMAR=?
Reference spec 3GPP TS 27.007.

Implementation for scenarios not explicitly covered in spec

1. Specs does not specify an option for SD card erase.
Implementation has an additional argument in CMAR command
for this.

2. Specs does not support pattern lock pin code
Phone will throw error if it is pattern locked.

3. Spec mandates CMAR pin_lock_code verification
If numeric/alphanumeric password is not set in phone ,
or if lock code is not provided in CMAR,
phone will throw error.

Change-Id: I32a05da767238a0440b01315d21b583a81278bc3