libstagefright: Fix for handling empty buffers from extractors
There is a possiblity that mSource->read will return a NULL buffer with
a invalid size.  As some implementations of memcpy don't handle NULL
buffers well, it's best to check for a NULL buffer and avoid the memcpy.

This fix also handles the case where the buffer happens to be NULL but
size happens to be wrongly set.

Change-Id: I52c7b11c01d20ef5bf6ebaa32a3b62359bffacaa