surfaceflinger: Avoid calling wormhole for fullscreen apps
For apps that use the full screen bypass,do not call drawWormhole
as this causes glClear commands to be accumulated in the GPU command
buffer, which causes a memory leak if it is called over and over
again without a corresponding glFinish(), glFlush() or eglSwapBuffers()
Earlier this was fixed by calling a glFinish() in drawWormhole
however that had the side effect of increasing power consumption
for full screen apps, since the GPU is also woken up even though
it is not necessary. This glFinish() is now removed.

Change-Id: Ia8101494430133e3ba7acf3b39644b6df9921fe4
CRs-fixed: 296504