surfaceflinger: Avoid calling wormhole for fullscreen apps
For apps that use the full screen bypass,do not call drawWormhole
as this causes glClear commands to be accumulated in the GPU command
buffer, which causes a memory leak if it is called over and over
again without a corresponding glFinish(), glFlush() or eglSwapBuffers()
Earlier this was fixed by calling a glFinish() in drawWormhole
however that had the side effect of increasing power consumption
for full screen apps, since the GPU is also woken up even though
it is not necessary. This glFinish() is now removed.

CRs-fixed: 296504
(cherry picked from commit 063289643a1e9ac474adade2570de40f9352c24b)

Change-Id: I2e18b96f9884702985b7e89c3e27eb521fe538e8