frameworks/base: sync to filesystem after setting wallpaper
This change prevents the livewallpaper from reverting to
the default wallpaper if the target crashes or is
otherwise reset abruptly after changing the wallpaper.
This was occurring because wallpaper_info.xml is of
filesize 0 while it is waiting to be synced.

CRs-Fixed: 294368
(cherry picked from commit c6803d9ca93b20f76d2d19eeb586e18ed52c29fd)

Change-Id: I9d633cd58ec2c95c9b70925ada318261da6e7139