Always accept OMX.TI.DUCATI.VIDDEC for H.264
In the QueryCodecs routine, OMX components are opened so their capabilities
can be queried, then closed.  The goal is to create a list of candidate codecs
for a certain mime type.

This change makes it so when looking for an H.264 decoder, always accept
OMX.TI.DUCATI.VIDDEC as a candidate.  You can't open it up to query what
profiles it supports and close it again, since teardown is not supported.

Fortunately, OMX.TI.DUCATI.VIDDEC supports all the profiles anyway.

This routine is not used by AwesomePlayer, but is used by the Flash player.