Bluetooth: Fix to FTP and OPP concurrency
Since the methods and members of SRM are static in ObexHelper Clase, when FTP and OPP(or any two obex profiles)
are operating on SRM and non SRM modes(over L2CAP) the second session overwrites the SRM satus of first one.
To avoid issue, members and methods of SRM are modified to instance based(an instance per session) to operate
independently.

CRs-fixed: 298776

Change-Id: I764e28258deb29d7702cd5608b0cffe6f662d574