Camera : Prevent dead lock occurence in preview buffer callback
When creating Raw heap in takepicture,we wait for preview heap to get
deallocated.It may cause a dead lock in preview buffer callback (PCB),
as PCB is trying to aquire a lock which was already with takepicture.
To prevent this, we ignore preview frame callbacks coming after
takePicture request , without waiting to aquire lock.
This way , we create raw heap after destryoing preview heap
and also prevent the occurence of dead lock.

Change-Id: Ice5e353968bd1bc5f8fc2de52e9127ed7039d0ef
CRs-fixed: 299732