libstagefright: Decrement the reference count of pmem in destructor
pmem reference count needs to be decremented to correctly
release and unmap pmem in the kernel driver. The refernce count
gets deceremented while freeing the node instance before destroying
component instance. But there is a possibility that this code may
not be executed if OMXCodec object is already destroyed.
Fixing it by clearing the reference in OMXCodec destructor.

Change-Id: I9258fcee7e16fe9afcdf322db69d83a450dbb8e5
CRs-Fixed: 301193