frameworks/base: surfaceflinger: fix for multiple conversions in copybit
CRs-fixed: 303334

fixes color conversion in copybit happening mutiple times.
Issue is observed when camera is using YV12 format.
In copybit for the sake of MDP hardware buffer is converted
to YUV420.After that copybit trying to set current format of
buffer as YUV420.But that is not actual mNativeBuffer(which
layerbuffer is holding). copybit/libagl modules creating copies
of NativeBuffer and the format settinng after conversion is
never getting reflected in actual NativeBuffer of LayerBuffer.

Because of this, if for some reason,SF composes all layers
again(assume there is no fresh postBuffer),but copybit will
convert the buffer again and corrupts frame.Const for mNativeBuffer
is removed to overwrite format of mNativeBuffer and using copybit
stretch as a inplace conversion routine.

Change-Id: Ib6531f444e307e1517aebfe6de2c35419005584f