libstagefright: Check for continuous NAL length errors while playback.
In MPEG4 parser, if we encounter a corrupt NAL we ignore the error
and continue parsing. If there are continuous NAL length errors
until the end of the clip, seek results in ANR. To handle this check
if there are continuous NAL Length errors for 10 frames and abort the
playback gracefully.

Change-Id: Ie59d8ed713fa996893c6513aa60b774efdcdd2d7
CRs-Fixed: 305756