libstagefright: Check for continuous NAL length errors while playback.
In MPEG4 parser, if we encounter a corrupt NAL we ignore the error
and continue parsing. If there are continuous NAL length errors
until the end of the clip, seek results in ANR. To handle this check
if there are continuous NAL Length errors for 10 frames and abort the
playback gracefully.

CRs-Fixed: 305756
(cherry picked from commit 548e7f6572690b8bc995d787ae099a7b032fdeaa)

Change-Id: I190962a2575c0353ee37b82a62ec971f4c8f5e53