sip: Have SipService look into the new connected value
Under some circumstances with Wifi
we are seeing the sequence of events:
    Connected
    Connecting/Authenticating
    Connected

SipService looks into the previous value of connected, but doesn't verify that
the current message is due to a disconnect. Because of that, the last Connected
causes SipService to assume that service was lost, and ends all sip calls.

This change checks for the current value and does nothing if both new and old
states are connected

Change-Id: Ib3b008e7ac92d5746fbb2a9aa7684d2e43be19ac
CRs-Fixed: 306245