libstagefright: Handle SHUTTING_DOWN state on O/P port
When component generates an event error, the port status is changed to
shutting down and error is returned to the frameworks. Stop called from
frameworks causes signalbufferreturned call in OMXCodec. In this function
there is an assert for O/P port status other than Pausing and Enabled. Added
Shutting down state to this CHECK condition

CRs-fixed: 304797
Change-Id: I82f141d7e9cd39859140d59212641c2b319c8a9c