libstagefright: In LPA seek update RequestQueue to handle correct buffer sequence.
-Issue: When LPA seek is done, buffers in RequestQueue
and ResponseQueue are not drained causing wrong buffer
sequence resulting in no audio.
-Fix: In LPAPlayerALSA, when seek is done buffers need
to be drained both in RequestQueue and ResponseQueue.
RequestQueue need to provide buffers from start of
the bufPool after seek. This is needed so the buffers
provided to kernel will be in the correct sequence.

CRs-Fixed: 311642
CRs-Fixed: 311653
CRs-Fixed: 311652
Change-Id: I8532548d520cf03b0e0dcc476cfb558f063cf608