VideoView: Show video controls correctly on pause
Currently, the video control panel is not updated if rotation occurs
when the video is paused. So, the video gets rotated but the controls
are still shown for the old orientation.

Change this behavior by calling show() on the controller whenever
there is a surfaceChanged callback, irrespective of the playback state
of the video.

CRs-fixed: 311852
(cherry picked from commit 2d78d09ac0a93462ae213a585c3ca07ffc21b9af)

Change-Id: I002fb4ebc90f3d784122de3ab1c8b2da1bd09a09