VideoView: Show video controls correctly on pause
Currently, the video control panel is not updated if rotation occurs
when the video is paused. So, the video gets rotated but the controls
are still shown for the old orientation.

Change this behavior by calling show() on the controller whenever
there is a surfaceChanged callback, irrespective of the playback state
of the video.

Change-Id: Ic9bb9d6bac89584b72eec0602320aefc38630192
CRs-fixed: 311852