Telephony: Re-work CatService clean up and handle ICC refresh
* CatService is never disposed, only the STK icon is cleared
when the card goes away, or card is not available.

* CatService, instead of UiccCard, determines when to clean the
STK icon based on the card state

* CatService is not disposed to also avoid the race condition
in the following scenario:
1. Radio is on and card is present
2. Turn Radio Off
3. Turn Radio On
4. Proactive command received

If CatService was disposed when card goes away on radio off
then it is possible for the unsol proactive command to be received
before Catservice is up and running on radio on.

CRs-Fixed: 283523

* When CatService is disposed, quit the thread that was created

* Handle modem power off and ICC refresh in UiccManager and Cat

When CatService receives radio unavailable or power off event it
sends that event to Stk app so that it can uninstall itself.

IdleModeText and SETUP_MENU need to be cleared when ICC status changes to
REFRESH_RESET or RESET_INIT. Hence sending the Icc Status change to STkApp

CRs-Fixed:270755

Change-Id: If0c86ec13a0f31bf1a48a268f3da9b7e7a41559a