Bluetooth: Merge frameworks/base changes of 'ics' into jb
frameworks/base: cleanup of BluetoothHeadset service while BT turnoff

Connection to BluetoothHeadset service isn't unbound when
bluetooth is disabled. This adds to the number of active
connections to bluetooth headset service and avoids
garbage collection of death monitors. Fix is to unbind
the service when bluetooth is disabled.

Change-Id: I04aa9de8729ebce664cdffe4659a4eaa30c2e455

Bluetooth: Start hcidump automatically.

- Start "hcidump -xtw /data/hcidump/{time_stamp}.cfa"
  if bluetooth.hcidump.enabled = true
- Redirect "hcidump -xt" to logcat

Change-Id: I6158e3dfcb6618a79c7329f25d8b8e8ee8dd67e5
CRs-fixed: 323983

Bluetooth: Resolve memory leak in MR1-release changes.

Connection to BluetoothHeadset service isn't unbound when BT is
disabled. This adds to the number of active connections to
BluetoothHeadset service and avoids garbage collection (GC).
Now calls to unbind BluetoothHeadset service and also GC for
service listener references are balanced in this client.

CRs-fixed: 328720
Change-Id: Ib4d1e0c268b3abc837b61bf13be2402c8e95bebd

Bluetooth: Remove resource leaking.

- The BluetoothDeviceProfileState is not removed after unpair.
  As a result, the thread that handles StateMachine is leaked.
  It is now removed after unpair properly.
- The clean up has not been excuted when state machine is quiting
  because clean up is only invoked quiting in BondedDevice state.

Change-Id: I92a86a534bfbc30ed792e519ea0afe4bfdc8d7e5

Conflicts:

	core/java/android/bluetooth/BluetoothDeviceProfileState.java

Bluetooth: Fix Bluetooth shutoff sequence

-During Bluetooth shutoff (Power Off) first bring down the
eventloop, then call the BluetoothService shutoff
-Without the above change, there could be a deadlock if
BluetoothService tries to access eventloop while eventloop
is in the process of shutdown.

Change-Id: I719a720a91094ce939a564eba51a7a84324d1d7d
CRs-fixed: 336675

frameworks: Add unique UID for BLUETOOTH user

Change-Id: I1b583aa31fd06b5f3bc7c733df3ab7e3241547bb

Bluetooth: BluetoothAdapter: Check the availability of MainLooper before using it

The change introduced in ICS for createNewRfcommSocketAndRecord
API to use the MainLooper does not take care of the case where
the MainLooper is not available. This issue is faced by the SppTest
application and the application crashes for this reason. This change
uses mainLooper if it is available for the caller, if not it uses
the default one.

Change-Id: I3a8bbbadf321ea4fc9637887e77e4bcc3401c791