Wifi: Support IPv6
Enable IPv6 support in Wifi State machine. Since IPv6 default route
addition by the kernel is disabled, the state machine issues a Router
Solicitation and looks for a Router Advertisement(RA). The source
address of the RA is used as the gateway and the router information is
propogated to Connectivity Service, which will then add the default
route.

Assumptions/Requirements for Wifi network interface
* Must be configured for IPv6
* Stateless auto config must be enabled

Change-Id: I3afee1fc40406e076e0fc9244f50c058464461d0
CRs-Fixed: 333609

Conflicts:

	core/java/android/os/INetworkManagementService.aidl

Wifi: Remove time duration assumption when acquiring wakelock for v6 Gateway

The state machine acquires a wakelock when the IPv6 gateway renewal timer
expires. The State machine makes an assumption of how much time the
renewal will take and sets this time when acquiring the wakelock, so
that it will be automatically released.

If the time it takes to acquire v6 exceeds this time, it results in a
Runtime exception, crashing the system process.

This change removes any assumptions made by the state machine

Change-Id: Iad8eda47ee9a0d40c67d1d3a97cfd73e41e96e6e
CRs-Fixed: 362789

Wifi: Compare entire address when updating link properties

This fixes the issue where the same IP address was being added into
the link property multiple times.

While updating the IPv6 gateway and addresses, the state machine
queries the IP addresses associated with the WLAN interface. It then
adds each address to the link property.

Fix the state machine to compare the entire IP address and not just
the network part to determine if the link property already contains
the address.

Change-Id: I026731fcedeaec74d68bf9197ea201dbc982e343
CRs-Fixed: 366832