Wifi: Support IPv6
Enable IPv6 support in Wifi State machine. Since IPv6 default route
addition by the kernel is disabled, the state machine issues a Router
Solicitation and looks for a Router Advertisement(RA). The source
address of the RA is used as the gateway and the router information is
propogated to Connectivity Service, which will then add the default
route.

Assumptions/Requirements for Wifi network interface
* Must be configured for IPv6
* Stateless auto config must be enabled

Change-Id: I3afee1fc40406e076e0fc9244f50c058464461d0
CRs-Fixed: 333609