Telephony: Trigger records to be read when switching from NV to RUIM
When we boot up in NV mode on a device that has RUIM as well,
IccRecords is valid. IccRecords dont change when we switch from
NV to RUIM. Explicitly trigger records to be read when we switch,
rather than relying only on IccRecords to change.

Also refactor how quiet mode is handled.

Change-Id: I5c460d8dc1885847f0c7deaff2bdbcf480fe9fe4