frameworks/base: move the traces file to avoid data loss
If an ANR comes just after the framework reboot, the traces for
the framework reboot will get over-written with the ANR traces.
So, We are missing the traces for the framework reboot which
makes debugging the framework reboot difficult.

This patch will rename this trace file for the framework reboot
once the ActivityManager collects the traces.

(cherry picked from commit ca662d0a3809decd00b5acaca55462b7f5c66800)

Change-Id: Ic2362f5c44c8974c452f802789ac3b16b581aa73