Telephony: Refactor IMSI and SubscriberId handling for CDMAPhone/CDMALTEPhone
IccRecords
CdmaServiceStateTracker does not need to set IMSI on an IccRecord.

CDMAPhone
When queried for the subscriber Id return the IMSI from CdmaServiceStateTracker
or the card depending on whether the active subscription source is NV or RUIM
respectively.

CDMALTEPhone
Don't return the IMSI from the SIM app (which is used for LTE) when queried
for the SubscriberId. Defer to the super class CDMAPhone's implementation.

RuimRecords
- Override the getIMSI method
- Remove our local copy of the IMSI
- Don't read EF_CSIM_IMSIM using SIM I/O to get the IMSI and MIN.
- Use the RIL API to get the IMSI for a given app.

CRs-Fixed: 323977,342146

Change-Id: Ie9bb9d5019f962e003668dfb4029ca33af63eb48