PM: Enable Touch Wake Up
Normally touch events do not wake up the device from sleep.
Modify frameworks/base to add wake flag to touch events
when the screen is off.
The flag is added in interceptMotionBeforeQueueing since
touch events come in the catagory of motion events.
This approach does not work for omap3evm as touch events
are not reported to userspace when display goes off.
Tested on am335xevm.

Ported from ICS, commit-ID:  d3e6665

Signed-off-by: Arun Joseph <arunjoseph@ti.com>
Signed-off-by: Umakanta Patro <umakanta.patro@ti.com>