Telephony: Allow use of BearerType in Dual-IP retry
This change accomdates the use of bearer type when configuring APNs when
Dual-IP partial retry is required. Before this change, there cannot be
multiple APNs with bearer type as the differentiator.

Important note:
Dual-IP partial failure retry will only occur using the same APN on
which the original data call was established. For e.g: if LTE-APN and
EHRPD-APN are both configured as the default APN and the UE originates
the initial call on LTE-APN and then goes through multiple IRAT
handovers, any partial failure retry will take place on the LTE-APN.

Change-Id: I009acfc2077769ee92cc71be8e5ba74ddd042b8f
CRs-Fixed: 343088