Telephony: Don't broadcast CARD_IO_ERROR when there are no card apps
On bootup, it is possible for the card state to be non-ABSENT but
the number of applications to be 0. This is not an error.

One scenario when this could happen is:

On bootup the phone type could be GSMPhone but there could be no
SIM/USIM application on the card. In this case, we should not
broadcast error.

Change-Id: Ie12d335d42d198d16b828047f7c6716ae897e74d