Bluetooth: BluetoothAdapter: Check the availability of MainLooper before using it
The change introduced in ICS for createNewRfcommSocketAndRecord
API to use the MainLooper does not take care of the case where
the MainLooper is not available. This issue is faced by the SppTest
application and the application crashes for this reason. This change
uses mainLooper if it is available for the caller, if not it uses
the default one.

Change-Id: I3a8bbbadf321ea4fc9637887e77e4bcc3401c791