VideoView: Added suspend-resume support for 3D clips
On suspend-resume, 3D clips were not being detected as 3D and were being
played on 2D.  As we never send the frame with 3D SEI (typically the first
frame) on resume, it is necessary to store state in between.  Due to the
loss of the suspend-resume APIs in stagefright, mediaplayer can no longer
store state between s/r.  As a result, we have to make VideoView hold state.

We send a notification (MEDIA_INFO_3D_ATTRIBUTE)to the upper layer
with the appropriate 3D flags.  On resume, the upper layer is expected
to pass these flags back to stagefright via the setParameters call. This
in turn sets the appropriate key in metadata. In processSEI, if we
happen to see this key, we pass the value associated with the key to
NativeWindow.

we set the param only if video playback resumes from suspend

Change-Id: I3f6e1a89b84e541e7b1f3091d739959f9b3d0144
CRs-Fixed: 355006, 347382