Telephony:Do not update UserPreferredSub on deactivate failure
RIL_REQUEST_SET_UICC_SUBSCRIPTION may get failure whenever the
modem is process of radio power off. Whenever the user tries to
toggle the airplane on and off, the subscription activation may
get failure due to this.
Do not update the user preferred subscription in this
cases to avoid unnecessary prompt for the new card dialog
box. Process the card informations in subscriptionManager only
when the radio is ON to avoid unnecessary activate requests
sent to modem and avoid accessing the card information when
radio is powered off.

Change-Id: If4e4ee92d00a1cd8af3a9a851d522b368f8f8a46
CRs-Fixed: 346970