Fix Navigation bar icons not showing for landscape lcd
For landscape displays like 800x400, the Navigation bar appears on
the right-side instead of at the bottom of the screen.

However, the icons (Back, Home etc) were not showing on this bar.
Fixed this as per suggestion from Vladimir Barinov in rowboat group
http://groups.google.com/group/rowboat/browse_thread/thread/94feb97b294a1e44

An new system property lcd.landscape is used to determine the layout of the
keys in navigation bar.
- If lcd.landscape is false, use the normal logic
- If lcd.landscape is true, rotate the icons by 90deg

The system property can be set in init.rc for the respective device as follows:
    setprop lcd.landscape true

Signed-off-by: Vishveshwar Bhat <vishveshwar.bhat@ti.com>