ActivityManager: Only call trim memory on Launcher when memory is low
Calling the launcher's onTrimMemory adds 20-40ms to the time it takes
to launch an app, also the resources released by the calling of
onTrimMemory will most likely be needed when the user switches back to
the launcher app. The appp launch latencies are improved by only calling
onTrimMemory on the Launcher when it is required, such as in low memory
conditions or when the Launcher has been in the background for a
substantial amount of time.

CRs-Fixed: 373773
(cherry picked from commit 0e6bc9b44f218fa47328dce141b17a2d3d414241)

Change-Id: Ice91f27d488b30c8271cde7bd91386081745b6d5