Bluetooth: Avoid wrong UI pop up incase of auto pairing failure.
With headset class of devices auto pairing will be attempted with 0000
as pin key. When it is rejected there will be bond retry with UI
pop up for PIN key. Showing the key error on UI in this case before
reattempt with UI pop up for PIN Key is not user friendly. The current
change avoids the UI pop for first failure.

CRs-Fixed: 361025

Change-Id: I289ee91bf0be3fc838ba3f5fb4a6353bb36b4308