Wifi: Remove time duration assumption when acquiring wakelock for v6 Gateway
The state machine acquires a wakelock when the IPv6 gateway renewal timer
expires. The State machine makes an assumption of how much time the
renewal will take and sets this time when acquiring the wakelock, so
that it will be automatically released.

If the time it takes to acquire v6 exceeds this time, it results in a
Runtime exception, crashing the system process.

This change removes any assumptions made by the state machine

Change-Id: Iad8eda47ee9a0d40c67d1d3a97cfd73e41e96e6e
CRs-Fixed: 362789
(cherry picked from commit e9d325306f71006c39ef18aa97f5e64700a6527f)