Wifi: Compare entire address when updating link properties
This fixes the issue where the same IP address was being added into
the link property multiple times.

While updating the IPv6 gateway and addresses, the state machine
queries the IP addresses associated with the WLAN interface. It then
adds each address to the link property.

Fix the state machine to compare the entire IP address and not just
the network part to determine if the link property already contains
the address.

Change-Id: I026731fcedeaec74d68bf9197ea201dbc982e343
CRs-Fixed: 366832