Telephony(DSDS): Fix null pointer exception during subs active/deactive
When the user tries to toggle the airplane on and off,
the subscription activation/deactivation mechanism resulting in
null pointer exception as modem is in process of moving to radio
off/on states.

To fix this null checks are added when the subscriptions is
being activated/de-activated.

CRs-Fixed: 367967
(cherry picked from commit 95848b5d207f770f56b42101407e1f4b708582bf)

Change-Id: I889d8d4d0e95ea0e43daf0279a1d429a09033948