GPS: Fixes a bug in Hybrid provider operation
The changes fixes an issue when a Hybrid provider session has
been started when GPS has not been turned on in the Settings.
This is done by making sure the Criteria update is sent to
ULP engine irrespective of the settings for the providers.
CRs-Fixed: 368775

Change-Id: Idf094b36612075c834d3c8b2e07e9609d007ecd7
(cherry picked from commit 88e3657d33ad708a1723429bea934d347efc7501)