Telephony: Try setup data call when the conditions are met in applyNewState
The data connection failed is not broadcast for the enable APN type
requests when the data is disabled from the settings menu.
Changes to try intiate the data call if the conditions are met whenever
updating the enabled/met flags of the ApnContext to make sure that the
proper data connection state is broadcast.

Change-Id: I93c78f5f6cd269c61cc0a9b7876ff73d57f2bc18
CRs-Fixed: 371400