Telephony: Try setup data call when the conditions are met in applyNewState
The data connection failed is not broadcast for the enable APN type
requests when the data is disabled from the settings menu.
Changes to try intiate the data call if the conditions are met whenever
updating the enabled/met flags of the ApnContext to make sure that the
proper data connection state is broadcast.

CRs-Fixed: 371400
(cherry picked from commit 5f6ef98113e1f5fd26720c0c666e30318940e6f9)

Change-Id: I18f752369efa26ae3467c6fb48b199daf4dc1739