Fix for wifi-Direct ON/OFF.
Wifi-Direct was not turning on, after immediate turn on/off of wifi and
wifi-direct sequence.The supplicant was unable to find wlan0 interface.
So as a safer side, made a check that supplicant is started when the
driver is actually loaded.

CRs-fixed: 375259

Change-Id: I3f9f1412608bdab3ab0e48df9a57aa2c76f19788
(cherry picked from commit 22e4b2c4f1a637525b201b4d71e25852b76da38c)