Fix for wifi-Direct ON/OFF.
Wifi-Direct was not turning on, after immediate turn on/off of wifi and
wifi-direct sequence.The supplicant was unable to find wlan0 interface.
So as a safer side, made a check that supplicant is started when the
driver is actually loaded.

CRs-fixed: 375259

(cherry picked from commit 22e4b2c4f1a637525b201b4d71e25852b76da38c)

Change-Id: I7853ffaa3feb644c7604fbb911bf6aa60778dd44