frameworks/base: Set p2p group idle timeout to zero.
Set p2p group idle timeout to zero when the p2p group is removed.
Its value is set to 2 sec when the group is started and when the
same device becomes GO, the previous entry of p2p group idle time
is used which is causing the p2p group to get removed.Hence reset
p2p_group_idle value when the group is removed.

Change-Id: I29fafc3bc1f87842d13c9c834a4585e1a7b5ec9e
CRs-fixed: 385479