Telephony(DSDS): Extend the MSim CDMA Phone class from CDMALTEPhone
Currently CDMAPhone will not query the data registration state and
the it assumes that the data is registered on the radio technology
mentioned in the voice registration state.
Changes to extend the MSim CDMA Phone class from the CDMALTEPhone
to make sure that the data registration state is queried explicitly
and data connection is setup on the EvDO technology.

Change-Id: I9d2e4377f7bc90d3593434e06919877f3af1425e
CRs-Fixed: 373273