Telephony(DSDS): process cards after subscription status changed
The subscription manager currently ignores the SIM_STATUS_CHANGED
events if the subscriptions are not changed.  But the user can remove/insert
the SIM cards while in an emergency call, which will trigger the usol
SIM_STATUS_CHANGED event, but will not trigger the unsol
SUBSCRIPTION_STATUS_CHANGED event in case of card removal.  The unsol
SUBSCRIPTION_STATUS_CHANGED with deactivated event will be received
only after ending the emergency call. The Subscription Manager has
to process the available cards once the unsol SUBSCRIPTION_STATUS_CHANGED
with deactivated is received.

Changes made to explicitly trigger to process the available cards on
deactivated indication so that the cards info is processes and activates the
user preferred subscriptions or notify the new cards availability.

Change-Id: Idad787f60901e4d984bc7105874bc633bcae7d83
CRs-Fixed: 387198