Bluetooth: Reverted the impl to avoid pairing with HID Mouse devices.
If a HID Mouse device is 2.1 device, then pairing is required as DUT
is supposed to save the Link key. So if pairing is avoided at the
frameworks, there will be connection issues with 2.1 HID devices.
The Bluez space code changes made sure there are no connection issues
with HID 2.0 devices.

CRs-Fixed: 385463
Change-Id: Ib2abdfe6b0296ea9b93cd0207cd39bf61b17a3e9