Telephony(DSDS): Avoid DDS Switch Failure
Multiple instances of GsmDataConnection objects are getting
created for single apncontext because of DDS switch between
Sub1 and Sub2. These are getting cleared (making to null)
irrespective of their state like in sync or not. This results
in telephony framework getting out of sync with modem and also
it block the subscription UI from changing it state and hence
the further subscription switch are not processed.

These should get cleared when they are in inactive state/out
of sync only.

Change-Id: I51d6777b58d715fce8f04bbdc69ccc8f0545f42f
CRs-fixed: 393427