webkit: Create HTML5VideoView only when there is a valid EGL context
Depending on when HTML5VideoView is created, it is possible that there
is currently no EGL context attached. In this case, the call to generate
a GL texture for the video SurfaceTexture will fail returning a bad
texture name. When this happens, video frames will fail to be rendered
resulting in black video frames.

To prevent this, validate the current EGL context before creating a new
HTML5VideoView. If validation fails, requeue the command to be tried
again at a later time.

Change-Id: I9098998e08a5be7c9680075a688df8d586c934ba
CRs-Fixed: 391616