libstagefright: Checks before freeing buffers on a port
OMXCodec calls dequeueBuffer on the NativeWindow in a loop
and registers output buffers with the OMX component.
If a dequeueBuffer fails, the buffers already registered
with the OMX component were not getting freed laeding
to a memory leak.
On an error, check state of OMXCodec before freeing buffers on a port.

Change-Id: I62d60d9a75d783ff834c247995ff5a88caacbfa1
CRs-Fixed: 393352