Bluetooth: HFP profile is not connected randomly when user repeatedly            turns On and Off.
The BluetoothAdapterStateMachine checks for current active ACL connections
if it is 0 it directly moves from BluetoothOn to PowerDown state but the
BluetoothHeadsetService state is in still Connecting state in this case.
Added one additional check here to make sure the HFP state is in
DISCONNECTED state before moving to PowerDown state.

CRs-Fixed: 394077

Change-Id: Id872ac1788cfa60398059390c6d86a6d75941a8f