Bluetooth: Fix "No such method on org.bluez.characteristic" D-bus error
When the LE gatt client is unbonded from the remote, the gatt
characteristics object paths are unregistered from D-bus but
not cleared correctly in the frameworks cache. When the gatt
client is bonded again, stale characteristic object path that
is not registered with D-bus is retrieved from the cache.
Added fix to clear the char object paths on unbonding.

CRs-fixed: 399535
Change-Id: I961108c75d92dca2d63d060f5feae255c75eee2b