Telephony: Do not broadcast card error intent when app index is invalid
IccCardProxy broadcasts card error intent when application index
received from UIM is invalid(i.e.mUiccAplication is null) during
initialization of card for fraction of time. As a result of which
"Invalid Card" is displayed on UI momentarily.

To fix this, When app index received from UIM is invalid broadcast
"NOT_READY" intent.

Change-Id: Ia00f1fde62df988d9bba69c5ae8f55a247580f16