Telephony: Use system property for default subscription
Some of the third party apps do not have access to the sqlite database
properties defined in telephony. As a result, a security exception is
thrown in TelephonyManager.java when the default subscription sqlite
property is accessed.

Use a system property for default subscription which can be accessed
by third party apps.

Change-Id: I9a6049563cbd003ba4a3c80da20c855c536b6062
CRs-Fixed: 397813