wifi: increment buffer size for wifi supplicant string command
The length of a string command sent to wpa_supplicant was limited to 256.
This limitation does not match the implementation in wpa_supplicant.
The supplicant can accept string commands in some cases up to 4096
bytes long and performs a command specific check for
a command length not to exceed maximum size.

Changed buffer size for string commands in frameworks to 4096 in order to
allow for longer commands (e.g., service request).

CRs-fixed: 402625

Change-Id: I6498bb57e0aeb166a1959d33b6285b2226c975c6