Telephony(DSDS): Trigger disable/enableApnType on both Sub's
When the user enable wifi, disableApnType will be called to
deactivate the data on current SUB1. User switch the DDS from
SUB1 to SUB2 and disable the wifi. In this case enableApnType
will be called to make data active on SUB2. Again user switch
the DDS from SUB2 to SUB1. Data is not becoming active on SUB1.

Changes made to call the disableApnType/enableApnType properly
to get the notification about both disable/enable data
connection on active and non-active DDS Sub's with DCT.

Change-Id: I15d5f3d8d3611263ed6c87a7fadba84175a20196
CRs-fixed: 406420