Telephony(DSDS): update IccRecords in MSim DCTs properly.
Currently the IccRecords object in the MSim DCTs is set on power up
or on the first DDS switch. So when the user switches the DDS before
the RECORDS LOADED event, only the DCT corresponds to the DDS will
get the event, and the DCT corresponds to the non-DDS will not receive.
The RECORDS LOADED event will never gets to the other DCT as it will
not register again for the records loaded event on DDS switch.

Changes made to set the IccRecords to null in the DCT for non-DDS on
dds switch, so that the DCT corresponds to the DDS will register again
for the records loaded event on the correct IccRecords object.
Changes in updateRecords in MSimCdmaDCT to update the IccRecords only
in case of DDS.

CRs-Fixed: 402570
(cherry picked from commit c3b40dc5279da6ffc7c9bc558cccb6a6edcb2d44)

Change-Id: I548343689bb2dce6ad47f569f9b48478c0515887