framworks/base: Do not enable configured networks on WPS events.
The frameworks , if enables the configured networks after the WPS
SUCCESS , the supplicant shall consider the networks based on the
priority for the WPA2 handshake after the successful WPS handshake.
This might end up in the station connecting to the AP other than
the current one with which the WPS handshake is done.Moreover,
since framework do not disable the networks and the wpa_supplicant
was the entity that disabled the networks before wps handshake , it
could actually be cleanest way to re-enable the networks that were
disabled during the WPS operation within wpa_supplicant.This patch
addresses this by not enabling the configured networks on WPS events.

Change-Id: Idc53c00fd339c7017564099455fd6f7d93370a1e
CRs-fixed: 398860