webkit: Abandon audio focus when HTML5 audio is paused
When audio pause is initiated by the audio element or when the
audio focus is lost for an undetermined amount of time, abandon
audio focus so that HTML5Audio will not trigger a resumption
of playback when the audio focus is regained.

Fixes undesired resuming of paused HTML5 audio after the user has
left the browser activity.

CRs-Fixed: 395215

Change-Id: I16e20171cad504694152950b62c3797b7a33e597