libstagefright:mpeg2ts: Do not store source reference in extractor
The reason for this change is because extractor holds a reference to
source and source holds one to extractor. So neither get destroyed
when AwesomePlayer releases its reference of both.

Without either the source and extractor getting destroyed, the file
fd is never closed, resulting in a leak.

Instead now the source will be created on the fly(as before), when
getTrack is called.

Change-Id: I993d95b46624f1b5c3a64593a77939b8a4c3eee9
(cherrypick from 6f1a424e94e6ddaa833b4d408dfe462d4b2adcb4)
CRs-Fixed: 414940