frameworks/base : Remove redundant isPaused variable setting.
-- LPA does not resume after EOS is reached while playing Dragon
Fight game.
-- In Normal Musicplayback, isPaused variable is set to true when
Pause is issued from Awesomeplayer and it is reset to false when
start/resume is called from Awesomeplayer .
-- isPaused is set to true even in the Event Thread when EOS is
reached. This is not needed as isPaused will not be reset to false
till start or resume is called from Awesomeplayer which blocks
decoder thread.
-- Removing the redundant code to set isPaused to true in Event
Thread when EOS is reached.

CRs-fixed: 431869
Change-Id: I5c568bf3b4c6b219c004020c64e5a0c8b882a135