Send user confirmation instead of Terminal Response for Setup Call
On a Setup Call proactive command, a confirmation dialog is displayed
to the user to confirm or reject the call. If the user hits the back
key to dismiss the dialog, consider that as the user rejecting the call,
instead of sending a Terminal Response.

CRs-fixed: 441003
(cherry picked from commit 5564edb8b5eef0f53fa887fb35aef3cb8338805a)

Change-Id: I5190bf4aed68e174a5894b0603fe986a948d3bb1