Bluetooth:Not doing updateDeviceServiceChannelCache addProperties during BT ON.
updatingDeviceServiceChannelCache is having a lot of dbus calls via native
and we call it for each bonded device before turning on Bluetooth.
So with multiple devices bonded. we were getting blocked on this infinitely,
hence  BT gets stucked in turning on. So now we are updating it after
broadcasting bluetooth on state.

Change-Id: I2f21b7c387f9d3b3bf26365191ecb994a9ee97d3