frameworks/base: disable FadingMarquee effect
The TextViews have fadingEdge effect on Marquee enabled. The fading edges
are created for labels when the text is longer than the Icon width, and the
fading edges are first drawn to the buffer and then hidden. The drawing of a
FadingEdge is very costly on GPU as it might cause a resolve - leading to poor
performance in the Launcher UX. This is especially the case in the text labels
of App Icons in the Hotseat area.

This change disables fadingEdge for Icons at a framework level. Instead of the
fading edge we now change the effect to ellipsis.

CRs-Fixed: 444778
Change-Id: I180fdeb78eb73c8ed2771578042735087b9f265b